requirements.txt
blessings==1.6
bpython==0.14.1
curtsies==0.1.19
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.8
django-classy-tags==0.6.1
django-cms==3.1.0
django-constance==1.0.1
django-extensions==1.5.2
django-mptt==0.6.1
django-reversion==1.8.5
django-sekizai==0.8.1
django-toolbelt==0.0.1
django-treebeard==3.0
djangocms-admin-style==0.2.5
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.1
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.7.1
djangocms-link==1.5
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.3
djangocms-video==0.1
djrill==1.3.0
ecdsa==0.13
Fabric==1.10.1
greenlet==0.4.5
gunicorn==19.2.1
html5lib==0.99999
ipdb==0.8
ipython==3.1.0
-e git+https://github.com/monoku/django-copy.git@6c85f8247e48a7dbab4096c8e1f0b48f8db3f01e#egg=kopy-master
paramiko==1.15.2
Pillow==2.7.0
psycopg2==2.5.4
pycrypto==2.6.1
Pygments==2.0.2
pytz==2014.10
redis==2.10.3
requests==2.5.1
six==1.9.0
South==1.0.2
static3==0.5.1
